v1.2-refactor: framework/render
