+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"/Users/vlad/Desktop/kottans-frontend/Free_time_app/cssnano.config.js":"1","/Users/vlad/Desktop/kottans-frontend/Free_time_app/index.js":"2","/Users/vlad/Desktop/kottans-frontend/Free_time_app/prettier.config.js":"3","/Users/vlad/Desktop/kottans-frontend/Free_time_app/.eslintrc.js":"4"},{"size":145,"mtime":1619102991855,"results":"5","hashOfConfig":"6"},{"size":1868,"mtime":1621878142546,"results":"7","hashOfConfig":"6"},{"size":192,"mtime":1619103003612,"results":"8","hashOfConfig":"6"},{"size":395,"mtime":1619091065275,"results":"9","hashOfConfig":"6"},{"filePath":"10","messages":"11","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"1th3zqe",{"filePath":"12","messages":"13","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"14","messages":"15","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"16","messages":"17","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/Users/vlad/Desktop/kottans-frontend/Free_time_app/cssnano.config.js",[],"/Users/vlad/Desktop/kottans-frontend/Free_time_app/index.js",[],"/Users/vlad/Desktop/kottans-frontend/Free_time_app/prettier.config.js",[],"/Users/vlad/Desktop/kottans-frontend/Free_time_app/.eslintrc.js",[]]
+[{"/Users/vlad/Desktop/kottans-frontend/Free_time_app/cssnano.config.js":"1","/Users/vlad/Desktop/kottans-frontend/Free_time_app/index.js":"2","/Users/vlad/Desktop/kottans-frontend/Free_time_app/prettier.config.js":"3","/Users/vlad/Desktop/kottans-frontend/Free_time_app/.eslintrc.js":"4"},{"size":145,"mtime":1619102991855,"results":"5","hashOfConfig":"6"},{"size":1909,"mtime":1621879246306,"results":"7","hashOfConfig":"6"},{"size":192,"mtime":1619103003612,"results":"8","hashOfConfig":"6"},{"size":395,"mtime":1619091065275,"results":"9","hashOfConfig":"6"},{"filePath":"10","messages":"11","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"1th3zqe",{"filePath":"12","messages":"13","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"14","messages":"15","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"16","messages":"17","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/Users/vlad/Desktop/kottans-frontend/Free_time_app/cssnano.config.js",[],"/Users/vlad/Desktop/kottans-frontend/Free_time_app/index.js",[],"/Users/vlad/Desktop/kottans-frontend/Free_time_app/prettier.config.js",[],"/Users/vlad/Desktop/kottans-frontend/Free_time_app/.eslintrc.js",[]]
