+++ src/components/GetNewActivity.js
@@ -1,0 +1,6 @@
+import renderActivity from './RenderActivity';
+
+export default function getNewActivity() {
+  window.dataStore.currentActivity = {};
+  renderActivity();
+}
