+++ src/data/dataStore.js
@@ -1,0 +1,6 @@
+const dataStore = {
+  currentType: '',
+  currentActivity: {},
+};
+
+export default dataStore;
