+++ src/data/types.js
@@ -1,0 +1,13 @@
+const types = [
+  'Education',
+  'Recreational',
+  'Social',
+  'DIY',
+  'Charity',
+  'Cooking',
+  'Relaxation',
+  'Music',
+  'Busywork',
+];
+
+export default types;
