+++ src/framework/render.js
@@ -1,0 +1,7 @@
+let Component, Target;
+
+export default function renderApp(componentFunction, targetElementId) {
+  if (componentFunction) Component = componentFunction;
+  if (targetElementId) Target = targetElementId;
+  document.getElementById(Target).innerHTML = `${Component()}`;
+}
